+++ inlang/source-code/message-bundle-component/src/stories/inlang-message-bundle.ts
@@ -2,9 +2,19 @@ import { html, LitElement } from "lit"
 import { customElement, property } from "lit/decorators.js"
 import { baseStyling } from "../styling/base.js"
 import overridePrimitiveColors from "../helper/overridePrimitiveColors.js"
-import type { MessageBundle, Message, Variant } from "@inlang/sdk/v2" // Import the types
+import type { MessageBundle, Message } from "@inlang/sdk/v2" // Import the types
 import { messageBundleStyling } from "./inlang-message-bundle.styles.js"
-import upsertVariant from "../helper/crud/variant/upsert.js"
+
+import "./inlang-variant.js"
+import "./inlang-lint-report-tip.js"
+
+import SlInput from "@shoelace-style/shoelace/dist/components/input/input.component.js"
+import SlButton from "@shoelace-style/shoelace/dist/components/button/button.component.js"
+import type { MessageLintReport } from "@inlang/sdk"
+
+// in case an app defines it's own set of shoelace components, prevent double registering
+if (!customElements.get("sl-input")) customElements.define("sl-input", SlInput)
+if (!customElements.get("sl-button")) customElements.define("sl-button", SlButton)
 
 @customElement("inlang-message-bundle")
 export default class InlangMessageBundle extends LitElement {
@@ -12,6 +22,9 @@ export default class InlangMessageBundle extends LitElement {
 
 	@property({ type: Object })
 	messageBundle: MessageBundle | undefined
+
+	@property({ type: Object })
+	lintReports: MessageLintReport[] | undefined
 
 	dispatchOnSetSettings(messageBundle: MessageBundle) {
 		const onChangeMessageBundle = new CustomEvent("change-message-bundle", {
@@ -40,106 +53,54 @@ export default class InlangMessageBundle extends LitElement {
 				<span># ${this.messageBundle?.id}</span>
 				<span class="alias">@${this.messageBundle?.alias.default}</span>
 			</div>
-			<div
-				@click=${() => {
-					// upsert variant
-					upsertVariant({
-						message: this.messageBundle!.messages[0]!,
-						variant: { match: ["other"], pattern: [{ type: "text", value: "Hello World!" }] },
-					})
-					this.requestUpdate()
-				}}
-			>
-				Add test variant
-			</div>
-			<div @click=${() => console.info(this.messageBundle)}>Log message bundle</div>
-			<div class="container">
-				${this.messageBundle?.messages.map((message) => this.renderVariantsTable(message))}
+			<div class="messages-container">
+				${this.messageBundle?.messages.map((message) =>
+					this._renderMessage(
+						message,
+						this.lintReports?.filter((report) => report.languageTag === message.locale)
+					)
+				)}
 			</div>
 		`
 	}
 
-	private renderVariantsTable(message: Message) {
-		// @ts-ignore
-		const selectors = message.selectors.map((selector) => selector.arg.name)
+	private _renderMessage(message: Message, messageLintReports?: MessageLintReport[]) {
 		return html`
-			<div class="variant-table">
-				<div class="lang">
+			<div class="message">
+				<div class="language-container">
 					<span>${message.locale}</span>
 				</div>
-				<table>
-					<thead>
-						<tr>
-							${selectors.map((selector) => html`<th>${selector}</th>`)}
-							<th>Pattern</th>
-							<th>Actions</th>
-						</tr>
-					</thead>
-					<tbody>
-						${message.variants.map((variant, index) =>
-							this.renderVariantRow(variant, index, selectors, message)
-						)}
-					</tbody>
-					<tfooter>
-						<td colspan="${selectors.length + 2}">Add Variant</td>
-					</tfooter>
-				</table>
-			</div>
-		`
-	}
-
-	private renderVariantRow(variant: Variant, index: number, selectors: string[], message: Message) {
-		const matches = selectors.map((selector) => {
-			const matchIndex = selectors.indexOf(selector)
-			return variant.match[matchIndex] || ""
-		})
-		return html`
-			<tr>
-				${matches.map((match) => html`<td>${match}</td>`)}
-				<td>
-					<input
-						type="text"
-						value=${variant.pattern
-							.map((p) => {
-								if ("value" in p) {
-									return p.value
-								}
-								return ""
-							})
-							.join(" ")}
-					/>
-					<div
-						@click=${(e: Event) => {
-							const target = e.target as HTMLInputElement
-							// upsert variant
-							upsertVariant({
-								message: message,
-								variant: {
-									match: variant.match,
-									pattern: [
+				<div class="message-body">
+					${message.selectors.length > 0
+						? html`<div class="message-header">
+								<div class="selector-container">
+									${message.selectors.map(
 										// @ts-ignore
-										{ type: "text", value: target!.previousSibling!.previousSibling!.value },
-									],
-								},
-							})
-							this._triggerSave()
-						}}
-					>
-						Save
+										(selector) => html`<div class="selector">${selector.arg.name}</div>`
+									)}
+								</div>
+								<div class="message-actions">
+									${messageLintReports && messageLintReports.length > 0
+										? html`<inlang-lint-report-tip
+												.lintReports=${messageLintReports}
+										  ></inlang-lint-report-tip>`
+										: ``}
+								</div>
+						  </div>`
+						: ``}
+					<div class="variants-container">
+						${message.variants.map(
+							(variant) =>
+								html`<inlang-variant
+									.variant=${variant}
+									.message=${message}
+									.triggerSave=${this._triggerSave}
+									.lintReports=${messageLintReports}
+								></inlang-variant>`
+						)}
 					</div>
-				</td>
-				<td>
-					<button
-						class="delete-button"
-						@click="${() => {
-							// eslint-disable-next-line no-console
-							console.log("delete variant", index)
-						}}"
-					>
-						🗑️
-					</button>
-				</td>
-			</tr>
+				</div>
+			</div>
 		`
 	}
 }