+++ inlang/source-code/paraglide/paraglide-unplugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inlang/paraglide-unplugin
 
+## 1.6.15
+
+### Patch Changes
+
+- Updated dependencies [2a5645c]
+  - @inlang/sdk@0.35.7
+  - @inlang/paraglide-js@1.9.1
+
 ## 1.6.14
 
 ### Patch Changes